Enahnce all the Existing APIs with GloablExceptionhandler and DTO's

+++ product-service/src/main/java/com/aeropelican/productservice/repository/ProductVariantsRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductVariantsRepository extends JpaRepository<ProductVariants, Integer> {
+public interface ProductVariantsRepository extends JpaRepository<ProductVariants, Long> {
 
 }
